Fixed config implementation in settings

+++ src/config.h
@@ -8,7 +8,7 @@ typedef char conf_address_t[256];
 typedef double coord_lat_t;
 typedef double coord_lon_t;
 typedef int conf_max_distance_t;
-typedef double conf_deviance_t;
+typedef int conf_deviance_t;
 
 typedef struct {
     conf_save_path_t shopping_list_save_path;
